Change Movie rate type from Float to Double

+++ creators_top1_of_Average_Rating.py
@@ -1,6 +1,6 @@
 from pyspark.sql import SparkSession
 from pyspark.sql import functions as F
-from pyspark.sql.types import StructType, StructField, StringType, FloatType
+from pyspark.sql.types import StructType, StructField, StringType, DoubleType
 from pyspark.sql.window import Window
 
 HDFS_INPUT_PATH = "hdfs://23133070-master:9000/du_lieu/tmdb_movies_normalized.parquet"
@@ -21,7 +21,7 @@ spark = SparkSession.builder \
 
 schema = StructType([
     StructField("Movie name", StringType(), True),
-    StructField("Movie rate", FloatType(), True),
+    StructField("Movie rate", DoubleType(), True),
     StructField("Description", StringType(), True),
     StructField("Creators", StringType(), True),
     StructField("Trailer link", StringType(), True),
@@ -65,3 +65,4 @@ final_df.write \
 print(f"--- Đã ghi kết quả vào MySQL bảng: {output_table} ---")
 
 spark.stop()
+
